[tracing] Fix types in SQL stdlib.
Also fix the Speedometer 3 SQL as it has mismatching type annotations,
but there annotations the correct and the logic is not.

Bug: b/353226722
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5705424
Commit-Queue: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1327507}

+++ base/tracing/stdlib/chrome/scroll_jank/scroll_offsets.sql
@@ -100,9 +100,9 @@ CREATE PERFETTO TABLE chrome_scroll_input_offsets(
   ts INT,
   -- The delta in raw coordinates between this scroll update event and the
   -- previous.
-  delta_y INT,
+  delta_y DOUBLE,
   -- The pixel offset of this scroll update event compared to the initial one.
-  relative_offset_y INT
+  relative_offset_y DOUBLE
 ) AS
 SELECT
   scroll_id,
@@ -132,9 +132,9 @@ CREATE PERFETTO TABLE chrome_presented_scroll_offsets(
   ts INT,
   -- The delta in raw coordinates between this scroll update event and the
   -- previous.
-  delta_y INT,
+  delta_y DOUBLE,
   -- The pixel offset of this scroll update event compared to the initial one.
-  relative_offset_y INT
+  relative_offset_y DOUBLE
 ) AS
 SELECT
   scroll_id,
